+++ packages/main/src/Input.ts
@@ -22,7 +22,6 @@ import {
 	isAndroid,
 } from "@ui5/webcomponents-base/dist/Device.js";
 import ValueState from "@ui5/webcomponents-base/dist/types/ValueState.js";
-import { getComponentFeature } from "@ui5/webcomponents-base/dist/FeaturesRegistry.js";
 import {
 	isUp,
 	isDown,
@@ -164,11 +163,6 @@ type InputSuggestionScrollEventDetail = {
  * When the user makes changes to the text, the change event is fired,
  * which enables you to react on any text change.
  *
- * **Note:** If you are using the `ui5-input` as a single npm module,
- * don't forget to import the `InputSuggestions` module from
- * "@ui5/webcomponents/dist/features/InputSuggestions.js"
- * to enable the suggestions functionality.
- *
  * ### Keyboard Handling
  * The `ui5-input` provides the following keyboard shortcuts:
  *
@@ -184,8 +178,6 @@ type InputSuggestionScrollEventDetail = {
  * ### ES6 Module Import
  *
  * `import "@ui5/webcomponents/dist/Input.js";`
- *
- * `import "@ui5/webcomponents/dist/features/InputSuggestions.js";` (optional - for input suggestions support)
  *
  * @constructor
  * @extends UI5Element
@@ -206,7 +198,6 @@ type InputSuggestionScrollEventDetail = {
 		ValueStateMessageCss,
 		SuggestionsCss,
 	],
-	features: ["InputSuggestions"],
 })
 
 /**
@@ -414,8 +405,6 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 	/**
 	 * Defines whether the component should show suggestions, if such are present.
 	 *
-	 * **Note:** You need to import the `InputSuggestions` module
-	 * from `"@ui5/webcomponents/dist/features/InputSuggestions.js"` to enable this functionality.
 	 * @default false
 	 * @public
 	 */
@@ -526,6 +515,12 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 	_accessibleLabelsRefTexts?: string;
 
 	/**
+	 * @private
+	 */
+	@property({ type: Object })
+	Suggestions?: InputSuggestions;
+
+	/**
 	 * Defines the suggestion items.
 	 *
 	 * **Note:** The suggestions would be displayed only if the `showSuggestions`
@@ -533,11 +528,6 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 	 *
 	 * **Note:** The `<ui5-suggestion-item>`, `<ui5-suggestion-item-group>` and `ui5-suggestion-item-custom` are recommended to be used as suggestion items.
 	 *
-	 * **Note:** Importing the Input Suggestions Support feature:
-	 *
-	 * `import "@ui5/webcomponents/dist/features/InputSuggestions.js";`
-	 *
-	 * automatically imports the `<ui5-suggestion-item>` and `<ui5-suggestion-item-group>` for your convenience.
 	 * @public
 	 */
 	@slot({ type: HTMLElement, "default": true })
@@ -583,7 +573,7 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 	_shouldAutocomplete?: boolean;
 	_keyDown?: boolean;
 	_isKeyNavigation?: boolean;
-	Suggestions?: InputSuggestions;
+	_indexOfSelectedItem: number;
 	_selectedText?: string;
 	_clearIconClicked?: boolean;
 	_focusedAfterClear: boolean;
@@ -649,6 +639,8 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 		this._isLatestValueFromSuggestions = false;
 
 		this._isChangeTriggeredBySuggestion = false;
+
+		this._indexOfSelectedItem = -1;
 
 		this._handleResizeBound = this._handleResize.bind(this);
 
@@ -728,6 +720,7 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 			const item = this._getFirstMatchingItem(value);
 			if (item) {
 				this._handleTypeAhead(item);
+				this._selectMatchingItem(item);
 			}
 		}
 	}
@@ -821,15 +814,22 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 		this._keyDown = false;
 	}
 
+	get currentItemIndex() {
+		const allItems = this.Suggestions?._getItems() as IInputSuggestionItemSelectable[];
+		const currentItem = allItems.find(item => { return item.selected || item.focused; });
+		const indexOfCurrentItem = currentItem ? allItems.indexOf(currentItem) : -1;
+		return indexOfCurrentItem;
+	}
+
 	_handleUp(e: KeyboardEvent) {
 		if (this.Suggestions?.isOpened()) {
-			this.Suggestions.onUp(e);
+			this.Suggestions.onUp(e, this.currentItemIndex);
 		}
 	}
 
 	_handleDown(e: KeyboardEvent) {
 		if (this.Suggestions?.isOpened()) {
-			this.Suggestions.onDown(e);
+			this.Suggestions.onDown(e, this.currentItemIndex);
 		}
 	}
 
@@ -1171,6 +1171,10 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 		this.Suggestions?.onItemPress(e);
 	}
 
+	_selectMatchingItem(item: IInputSuggestionItemSelectable) {
+		item.selected = true;
+	}
+
 	_handleTypeAhead(item: IInputSuggestionItemSelectable) {
 		const value = item.text ? item.text : "";
 
@@ -1212,10 +1216,12 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 			this.focused = false;
 		}
 
-		if (this._changeToBeFired) {
+		if (this._changeToBeFired && !this._isChangeTriggeredBySuggestion) {
 			this.fireDecoratorEvent(INPUT_EVENTS.CHANGE);
-			this._changeToBeFired = false;
+		} else {
+			this._isChangeTriggeredBySuggestion = false;
 		}
+		this._changeToBeFired = false;
 
 		this.open = false;
 		this.isTyping = false;
@@ -1258,10 +1264,19 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 			return;
 		}
 
-		const Suggestions = getComponentFeature<typeof InputSuggestions>("InputSuggestions");
-		if (Suggestions) {
+		const setup = (Suggestions: typeof InputSuggestions) => {
 			Suggestions.i18nBundle = Input.i18nBundle;
 			this.Suggestions = new Suggestions(this, "suggestionItems", true, false);
+		};
+
+		// If the feature is preloaded (the user manually imported InputSuggestions.js), it is already available on the constructor
+		if (Input.SuggestionsClass) {
+			setup(Input.SuggestionsClass);
+		// If feature is not preloaded, load it dynamically
+		} else {
+			import("./features/InputSuggestions.js").then(SuggestionsModule => {
+				setup(SuggestionsModule.default);
+			});
 		}
 	}
 
@@ -1404,7 +1419,6 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 		if (shouldFireSelectionChange) {
 			this.fireSelectionChange(suggestionItem, true);
 		}
-
 		this.acceptSuggestion(suggestionItem, keyboardUsed);
 	}
 
@@ -1735,16 +1749,9 @@ class Input extends UI5Element implements SuggestionComponent, IFormInputElement
 
 		return value;
 	}
-}
 
-// declare module "@ui5/webcomponents-base/jsx-runtime" {
-// 	// eslint-disable-next-line @typescript-eslint/no-namespace
-// 	namespace JSX {
-// 		interface IntrinsicElements {
-// 			"ui5-input": Input["_jsxProps"];
-// 		}
-// 	}
-// }
+	static SuggestionsClass?: typeof InputSuggestions;
+}
 
 Input.define();